// src/types/patient.d.ts
export interface Patient {
  id: string;
  first_name: string;
  last_name: string;
  dob?: string;
  phone?: string;
  email?: string;
  address?: string;
  city?: string;
  state?: string;
  zip?: string;
  medical_conditions?: string[];
  allergies?: string[];
  medications?: string[];
  emergency_contact_name?: string;
  emergency_contact_phone?: string;
  created_at: string;
  updated_at: string;
  gender?: string;
  primary_insurance?: string;
  secondary_insurance?: string;
  legacy_display_id?: string;
  marital_status?: string;
  occupation?: string;
  preferred_language?: string;
  blood_type?: string;
  height?: string;
  weight?: string;
  primary_care_physician?: string;
  last_physical?: string;
  usual_transport_mode?: string;
}